Fixed bug with saving last DataSet. Added test for this case.
