+++ src/main/java/ru/naumen/perfhouse/parser/Storage.java
@@ -3,7 +3,7 @@ package ru.naumen.perfhouse.parser;
 import org.influxdb.dto.BatchPoints;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import ru.naumen.perfhouse.influx.InfluxDAOInterface;
+import ru.naumen.perfhouse.influx.InfluxDAO;
 
 @Service
 class Storage {
@@ -12,24 +12,24 @@ class Storage {
     private long currentKey;
     private DataSet currentSet;
     private BatchPoints points;
-    private InfluxDAOInterface influxDAO;
+    private InfluxDAO influxDAO;
 
     @Autowired
-    Storage(InfluxDAOInterface influxDAOInterface) {
-        influxDAO = influxDAOInterface;
+    Storage(InfluxDAO influxDAO) {
+        this.influxDAO = influxDAO;
     }
 
     void init(String dbName, boolean printLog) {
         currentDb = dbName.replaceAll("-", "_");
         log = printLog;
 
+        influxDAO.connectToDB(currentDb);
+        points = influxDAO.startBatchPoints(currentDb);
+
         if (log)
         {
             System.out.print("Timestamp;Actions;Min;Mean;Stddev;50%%;95%%;99%%;99.9%%;Max;Errors\n");
         }
-
-        influxDAO.connectToDB(currentDb);
-        points = influxDAO.startBatchPoints(currentDb);
     }
 
     DataSet get(long key) {
@@ -82,6 +82,7 @@ class Storage {
     }
 
     void save() {
+        store();
         influxDAO.writeBatch(points);
     }
 }
